+++ Editor/ConversionSystem/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+#if ENABLE_NAVIGATION_PACKAGE_RELEASE_FEATURES
+using System.Runtime.CompilerServices;
+[assembly: InternalsVisibleTo("Unity.AI.Navigation.Updater")]
+#endif