+++ Editor/NavMeshModifierVolumeEditor.cs
@@ -32,14 +32,17 @@ namespace Unity.AI.Navigation.Editor
             m_Area = serializedObject.FindProperty("m_Area");
             m_Center = serializedObject.FindProperty("m_Center");
             m_Size = serializedObject.FindProperty("m_Size");
-
+#if !ENABLE_NAVIGATION_PACKAGE_RELEASE_FEATURES
             NavMeshVisualizationSettings.showNavigation++;
+#endif
         }
 
+#if !ENABLE_NAVIGATION_PACKAGE_RELEASE_FEATURES
         void OnDisable()
         {
             NavMeshVisualizationSettings.showNavigation--;
         }
+#endif
 
         Bounds GetBounds()
         {
@@ -63,8 +66,12 @@ namespace Unity.AI.Navigation.Editor
 
             serializedObject.ApplyModifiedProperties();
         }
-
+        
+#if !ENABLE_NAVIGATION_PACKAGE_RELEASE_FEATURES
         [DrawGizmo(GizmoType.Selected | GizmoType.Active)]
+#else
+        [DrawGizmo(GizmoType.InSelectionHierarchy | GizmoType.Active)]
+#endif
         static void RenderBoxGizmo(NavMeshModifierVolume navModifier, GizmoType gizmoType)
         {
             var color = navModifier.enabled ? s_HandleColor : s_HandleColorDisabled;
@@ -90,7 +97,9 @@ namespace Unity.AI.Navigation.Editor
         [DrawGizmo(GizmoType.NotInSelectionHierarchy | GizmoType.Pickable)]
         static void RenderBoxGizmoNotSelected(NavMeshModifierVolume navModifier, GizmoType gizmoType)
         {
+#if !ENABLE_NAVIGATION_PACKAGE_RELEASE_FEATURES
             if (NavMeshVisualizationSettings.showNavigation > 0)
+#endif
             {
                 var color = navModifier.enabled ? s_HandleColor : s_HandleColorDisabled;
                 var oldColor = Gizmos.color;