+++ Editor/Updater/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+#if ENABLE_NAVIGATION_PACKAGE_RELEASE_FEATURES
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Unity.AI.Navigation.Editor.Tests")]
+#endif