+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0-pre.1] - 2022-04-27
+### Added
+* NavMeshSurface supports links generation.
+* NavMeshSurface supports HeightMesh baking.
+* New package Navigation window adapting the obsolete Unity Navigation window functionalities to the package workflow.
+
+### Changed
+* NavMeshSurface is using the _Background Tasks_ window to report the baking progress
+* Minimum supported version is increased to Unity 2022.2
+
 ## [1.0.0-exp.4] - 2021-07-19
 ### Changed
 * Documentation updated with changes from Unity manual